# frontend/packages/api-client/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "TripsAgent.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost/"
    }
  ],
  "paths": {
...
    "/api/v1/storefront/site": {
      "get": {
        "tags": [
          "Storefront"
        ],
        "operationId": "GetSite",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Storefront"
        ],
        "operationId": "CreateSite",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSiteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/storefront/site/settings": {
      "put": {
        "tags": [
          "Storefront"
        ],
        "operationId": "UpdateSiteSettings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/storefront/theme": {
      "get": {
        "tags": [
          "Storefront"
        ],
        "operationId": "GetSiteTheme",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteThemeResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Storefront"
        ],
        "operationId": "SaveSiteTheme",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteThemeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteThemeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/storefront/pages/{pageId}": {
      "get": {
        "tags": [
          "Storefront"
        ],
        "operationId": "GetSitePage",
        "parameters": [
          {
            "name": "pageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SitePageResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Storefront"
        ],
        "operationId": "SaveSitePage",
        "parameters": [
          {
            "name": "pageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveSitePageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SitePageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Storefront"
        ],
        "operationId": "DeleteSitePage",
        "parameters": [
          {
            "name": "pageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/storefront/pages": {
      "post": {
        "tags": [
          "Storefront"
        ],
        "operationId": "CreateSitePage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSitePageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SitePageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/storefront/versions": {
      "get": {
        "tags": [
          "Storefront"
        ],
        "operationId": "ListSiteVersions",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SiteVersionResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/storefront/versions/stage": {
      "post": {
        "tags": [
          "Storefront"
        ],
        "operationId": "StageSite",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteVersionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/storefront/versions/{versionId}/publish": {
      "post": {
        "tags": [
          "Storefront"
        ],
        "operationId": "PublishSiteVersion",
        "parameters": [
          {
            "name": "versionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteVersionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/storefront/versions/{versionId}/rollback": {
      "post": {
        "tags": [
          "Storefront"
        ],
        "operationId": "RollBackSiteVersion",
        "parameters": [
          {
            "name": "versionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteVersionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/storefront/preview-links": {
      "post": {
        "tags": [
          "Storefront"
        ],
        "operationId": "CreateSitePreviewLink",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SitePreviewLinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SitePreviewLinkResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateSitePageRequest": {
        "required": [
          "title",
          "slug"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "slug": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "CreateSiteRequest": {
        "required": [
          "templateCode"
        ],
        "type": "object",
        "properties": {
          "templateCode": {
            "type": "string"
          }
        }
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SitePageResponse": {
        "required": [
          "id",
          "slug",
          "pageType",
          "title",
          "isSystem",
          "showInNav",
          "position",
          "metaTitle",
          "metaDescription",
          "revision",
          "blocks"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string"
          },
          "pageType": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "isSystem": {
            "type": "boolean"
          },
          "showInNav": {
            "type": "boolean"
          },
          "position": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "metaTitle": {
            "type": [
              "null",
              "string"
            ]
          },
          "metaDescription": {
            "type": [
              "null",
              "string"
            ]
          },
          "revision": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "blocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SitePageBlockResponse"
            }
          }
        }
      },
...
      "SitePreviewLinkRequest": {
        "required": [
          "versionId"
        ],
        "type": "object",
        "properties": {
          "versionId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
...
      "SiteResponse": {
        "required": [
          "id",
          "status",
          "templateCode",
          "templateName",
          "settings",
          "pages",
          "staged",
          "published",
          "hasUnstagedChanges",
          "hasUnpublishedChanges",
          "publishCheck",
          "primaryHostname",
          "siteUrl"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "templateCode": {
            "type": "string"
          },
          "templateName": {
            "type": "string"
          },
          "settings": {
            "$ref": "#/components/schemas/SiteSettingsResponse"
          },
          "pages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SitePageSummaryResponse"
            }
          },
          "staged": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SiteVersionResponse"
              }
            ]
          },
          "published": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SiteVersionResponse"
              }
            ]
          },
          "hasUnstagedChanges": {
            "type": "boolean"
          },
          "hasUnpublishedChanges": {
            "type": "boolean"
          },
          "publishCheck": {
            "$ref": "#/components/schemas/SitePublishCheckResponse"
          },
          "primaryHostname": {
            "type": [
              "null",
              "string"
            ]
          },
          "siteUrl": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "SiteSettingsRequest": {
        "required": [
          "name",
          "seoTitle",
          "seoDescription",
          "flightSearchEnabled",
          "contactEmail",
          "contactPhone",
          "whatsAppNumber",
          "contactAddress",
          "socialLinks"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "seoTitle": {
            "type": [
              "null",
              "string"
            ]
          },
          "seoDescription": {
            "type": [
              "null",
              "string"
            ]
          },
          "flightSearchEnabled": {
            "type": "boolean"
          },
          "contactEmail": {
            "type": [
              "null",
              "string"
            ]
          },
          "contactPhone": {
            "type": [
              "null",
              "string"
            ]
          },
          "whatsAppNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "contactAddress": {
            "type": [
              "null",
              "string"
            ]
          },
          "socialLinks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SiteSocialLinkDto"
            }
          }
        }
      },
...
      "SiteThemeRequest": {
        "required": [
          "logoAssetId",
          "primaryColor",
          "secondaryColor"
        ],
        "type": "object",
        "properties": {
          "logoAssetId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "primaryColor": {
            "type": "string"
          },
          "secondaryColor": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "SiteThemeResponse": {
        "required": [
          "logoAssetId",
          "logoPreviewUrl",
          "primaryColor",
          "secondaryColor"
        ],
        "type": "object",
        "properties": {
          "logoAssetId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "logoPreviewUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "primaryColor": {
            "type": "string"
          },
          "secondaryColor": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    }
  },
...
}